requirements.txt
fastapi
uvicorn
pillow
numpy
jinja2
